requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.2.0
opencv-python-headless==4.9.0.80